Move to our global system from AsyncLocalStorage for the current task context storage
