+++ packages/core/src/v3/taskContext/types.ts
@@ -1,0 +1,6 @@
+import { BackgroundWorkerProperties, TaskRunContext } from "../schemas";
+
+export type TaskContext = {
+  ctx: TaskRunContext;
+  worker: BackgroundWorkerProperties;
+};
