+++ packages/core/src/v3/workers/index.ts
@@ -1,5 +1,4 @@
 export { TaskExecutor, type TaskExecutorOptions } from "./taskExecutor";
-export { taskContextManager, TaskContextSpanProcessor } from "../tasks/taskContextManager";
 export type { RuntimeManager } from "../runtime/manager";
 export { PreciseWallClock as DurableClock } from "../clock/preciseWallClock";
 export { getEnvVar } from "../utils/getEnv";
@@ -7,3 +6,4 @@ export { OtelTaskLogger, logLevels } from "../logger/taskLogger";
 export { ConsoleInterceptor } from "../consoleInterceptor";
 export { TracingSDK, type TracingDiagnosticLogLevel, recordSpanException } from "../otel";
 export { StandardTaskCatalog } from "../task-catalog/standardTaskCatalog";
+export { TaskContextSpanProcessor, TaskContextLogProcessor } from "../taskContext/otelProcessors";
